Upgrade nats plugin to expand support to 2.16.5 (#711)
