+++ .github/workflows/plugins-test.3.yaml
@@ -67,7 +67,7 @@ jobs:
           - vertx-web-3.6plus-scenario
           - mariadb-scenario
           - micronaut-http-scenario
-          - nats-2.14.x-2.15.x-scenario
+          - nats-2.14.x-2.16.5-scenario
           - quasar-scenario
           - baidu-brpc-scenario
           - baidu-brpc-3.x-scenario
