+++ CHANGES.md
@@ -5,6 +5,7 @@ Release Notes.
 9.4.0
 ------------------
 
+* Upgrade nats plugin to support 2.16.5
 
 
 All issues and pull requests are [here](https://github.com/apache/skywalking/milestone/222?closed=1)
