+++ apm-sniffer/apm-sdk-plugin/nats-2.14.x-2.16.5-plugin/src/main/java/org/apache/skywalking/apm/plugin/nats/client/define/NatsSubscriptionInstrumentation.java
@@ -21,19 +21,17 @@ import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.matcher.ElementMatcher;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.ConstructorInterceptPoint;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.InstanceMethodsInterceptPoint;
-import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.ClassInstanceMethodsEnhancePluginDefine;
 import org.apache.skywalking.apm.agent.core.plugin.match.ClassMatch;
 
 import static net.bytebuddy.matcher.ElementMatchers.named;
 import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 import static org.apache.skywalking.apm.agent.core.plugin.match.NameMatch.byName;
 
-public class NatsConnectionInstrumentation extends ClassInstanceMethodsEnhancePluginDefine {
+public class NatsSubscriptionInstrumentation extends AbstractWitnessInstrumentation {
 
-    private static final String ENHANCE_CLASS = "io.nats.client.impl.NatsConnection";
-
-    private static final String CREATE_DISPATCHER_INTERCEPTOR = "org.apache.skywalking.apm.plugin.nats.client.CreateDispatcherInterceptor";
-    private static final String REPLY_MSG_INTERCEPTOR = "org.apache.skywalking.apm.plugin.nats.client.DeliverReplyInterceptor";
+    private static final String ENHANCE_CLASS = "io.nats.client.impl.NatsSubscription";
+    private static final String NEXT_MSG_INTERCEPTOR = "org.apache.skywalking.apm.plugin.nats.client.SubscriptionNextMsgInterceptor";
+    private static final String CONSTRUCTOR_INTERCEPTOR = "org.apache.skywalking.apm.plugin.nats.client.NatsSubscriptionConstructorInterceptor";
 
     @Override
     protected ClassMatch enhanceClass() {
@@ -42,7 +40,19 @@ public class NatsConnectionInstrumentation extends ClassInstanceMethodsEnhancePl
 
     @Override
     public ConstructorInterceptPoint[] getConstructorsInterceptPoints() {
-        return new ConstructorInterceptPoint[0];
+        return new ConstructorInterceptPoint[] {
+            new ConstructorInterceptPoint() {
+                @Override
+                public ElementMatcher<MethodDescription> getConstructorMatcher() {
+                    return takesArguments(5);
+                }
+
+                @Override
+                public String getConstructorInterceptor() {
+                    return CONSTRUCTOR_INTERCEPTOR;
+                }
+            }
+        };
     }
 
     @Override
@@ -51,28 +61,12 @@ public class NatsConnectionInstrumentation extends ClassInstanceMethodsEnhancePl
                 new InstanceMethodsInterceptPoint() {
                     @Override
                     public ElementMatcher<MethodDescription> getMethodsMatcher() {
-                        return named("createDispatcher").and(takesArguments(1));
+                        return named("nextMessageInternal").and(takesArguments(1));
                     }
 
                     @Override
                     public String getMethodsInterceptor() {
-                        return CREATE_DISPATCHER_INTERCEPTOR;
-                    }
-
-                    @Override
-                    public boolean isOverrideArgs() {
-                        return true;
-                    }
-                },
-                new InstanceMethodsInterceptPoint() {
-                    @Override
-                    public ElementMatcher<MethodDescription> getMethodsMatcher() {
-                        return named("deliverReply").and(takesArguments(1));
-                    }
-
-                    @Override
-                    public String getMethodsInterceptor() {
-                        return REPLY_MSG_INTERCEPTOR;
+                        return NEXT_MSG_INTERCEPTOR;
                     }
 
                     @Override
