+++ docs/en/setup/service-agent/java-agent/Plugin-list.md
@@ -162,7 +162,7 @@
 - hutool-http-5.x
 - micronaut-http-client-3.2.x-3.6.x
 - micronaut-http-server-3.2.x-3.6.x
-- nats-client-2.14.x-2.15.x
+- nats-client-2.14.x-2.16.5
 - impala-jdbc-2.6.x
 - jdk-forkjoinpool-plugin
 - jetty-thread-pool
