+++ test/plugin/scenarios/nats-2.14.x-2.16.5-scenario/configuration.yml
