+++ test/plugin/scenarios/nats-2.14.x-2.16.5-scenario/support-version.list
@@ -14,5 +14,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+2.16.5
 2.15.6
 2.14.2
